feat: split resolver
