+++ packages/backend/server/src/core/workspaces/resolvers/index.ts
@@ -1,4 +1,5 @@
 export * from './blob';
 export * from './history';
 export * from './page';
+export * from './team';
 export * from './workspace';
