+++ packages/backend/server/src/core/workspaces/resolvers/workspace.ts
@@ -22,7 +22,6 @@ import {
   InternalServerError,
   MailService,
   MemberQuotaExceeded,
-  NotInSpace,
   RequestMutex,
   SpaceAccessDenied,
   SpaceNotFound,
@@ -33,7 +32,7 @@ import {
 import { CurrentUser, Public } from '../../auth';
 import { type Editor, PgWorkspaceDocStorageAdapter } from '../../doc';
 import { DocContentService } from '../../doc-renderer';
-import { FeatureManagementService, FeatureType } from '../../features';
+import { FeatureManagementService } from '../../features';
 import { Permission, PermissionService } from '../../permission';
 import { QuotaManagementService, QuotaQueryType } from '../../quota';
 import { WorkspaceBlobStorage } from '../../storage';
@@ -41,8 +40,6 @@ import { UserService, UserType } from '../../user';
 import {
   InvitationType,
   InviteUserType,
-  TeamWorkspaceConfigType,
-  UpdateTeamWorkspaceConfigInput,
   UpdateWorkspaceInput,
   WorkspaceType,
 } from '../types';
@@ -344,32 +341,6 @@ export class WorkspaceResolver {
     });
   }
 
-  @ResolveField(() => TeamWorkspaceConfigType, {
-    description: 'Team workspace config',
-    complexity: 2,
-  })
-  async teamConfig(@Parent() workspace: WorkspaceType) {
-    return this.feature.getWorkspaceConfig(
-      workspace.id,
-      FeatureType.TeamWorkspace
-    );
-  }
-
-  @Mutation(() => Boolean)
-  async updateWorkspaceTeamConfig(
-    @CurrentUser() user: CurrentUser,
-    @Args({ name: 'input', type: () => UpdateTeamWorkspaceConfigInput })
-    { id, ...configs }: UpdateTeamWorkspaceConfigInput
-  ) {
-    await this.permissions.checkWorkspace(id, user.id, Permission.Owner);
-
-    return this.feature.updateWorkspaceConfig(
-      id,
-      FeatureType.TeamWorkspace,
-      configs
-    );
-  }
-
   @Mutation(() => Boolean)
   async deleteWorkspace(
     @CurrentUser() user: CurrentUser,
@@ -486,47 +457,6 @@ export class WorkspaceResolver {
     }
   }
 
-  @Mutation(() => String)
-  async grant(
-    @CurrentUser() user: CurrentUser,
-    @Args('workspaceId') workspaceId: string,
-    @Args('userId') userId: string,
-    @Args('permission', { type: () => Permission }) permission: Permission
-  ) {
-    await this.permissions.checkWorkspace(
-      workspaceId,
-      user.id,
-      Permission.Owner
-    );
-
-    if (permission === Permission.Owner) {
-      throw new CantChangeSpaceOwner();
-    }
-
-    try {
-      // lock to prevent concurrent invite and grant
-      const lockFlag = `invite:${workspaceId}`;
-      await using lock = await this.mutex.lock(lockFlag);
-      if (!lock) {
-        return new TooManyRequest();
-      }
-
-      const isMember = await this.permissions.isWorkspaceMember(
-        workspaceId,
-        userId,
-        Permission.Read
-      );
-      if (isMember) {
-        return await this.permissions.grant(workspaceId, userId, permission);
-      } else {
-        return new NotInSpace({ spaceId: workspaceId });
-      }
-    } catch (e) {
-      this.logger.error('failed to invite user', e);
-      return new TooManyRequest();
-    }
-  }
-
   @Throttle('strict')
   @Public()
   @Query(() => InvitationType, {
@@ -628,27 +558,6 @@ export class WorkspaceResolver {
         inviteeName: invitee.name,
         workspaceName: workspace.name,
       });
-    }
-
-    return this.permissions.acceptWorkspaceInvitation(inviteId, workspaceId);
-  }
-
-  @Mutation(() => Boolean)
-  async declineInviteById(
-    @CurrentUser() user: CurrentUser,
-    @Args('workspaceId') workspaceId: string,
-    @Args('inviteId') inviteId: string
-  ) {
-    await this.permissions.checkWorkspace(
-      workspaceId,
-      user.id,
-      Permission.Admin
-    );
-
-    const { invitee, user: inviter } = await this.getInviteInfo(inviteId);
-
-    if (!inviter || !invitee) {
-      throw new UserNotFound();
     }
 
     return this.permissions.acceptWorkspaceInvitation(inviteId, workspaceId);
